+++ internal/pkg/config/root.go
@@ -8,10 +8,10 @@ package config
 
 import (
 	"bytes"
-	"fmt"
 	"net"
 	"os"
 	"reflect"
+	"strconv"
 
 	"github.com/creasty/defaults"
 	"github.com/warewulf/warewulf/internal/pkg/wwlog"
@@ -26,11 +26,11 @@ var cachedConf WarewulfYaml
 type WarewulfYaml struct {
 	Comment     string        `yaml:"comment,omitempty"`
 	Ipaddr      string        `yaml:"ipaddr,omitempty"`
-	Ipaddr6     string        `yaml:"ipaddr6,omitempty"`
 	Netmask     string        `yaml:"netmask,omitempty"`
 	Network     string        `yaml:"network,omitempty"`
-	IpCIDR6     string        `yaml:"-"`
 	Fqdn        string        `yaml:"fqdn,omitempty"`
+	Ipaddr6     string        `yaml:"ipaddr6,omitempty"`
+	PrefixLen6  string        `yaml:"prefixlen6,omitempty"`
 	Warewulf    *WarewulfConf `yaml:"warewulf,omitempty"`
 	API         *APIConf      `yaml:"api,omitempty"`
 	DHCP        *DHCPConf     `yaml:"dhcp,omitempty"`
@@ -140,16 +140,17 @@ func (conf *WarewulfYaml) Parse(data []byte, autodetect bool) error {
 		}
 	}
 
-	if conf.Ipaddr6 != "" {
-		if ip, _, err := net.ParseCIDR(conf.Ipaddr6); err == nil {
-			if ip.To4() != nil {
-				return fmt.Errorf("invalid ipv6 address: ip address is ipv4: %s", conf.Ipaddr6)
-			}
-			conf.IpCIDR6 = conf.Ipaddr6
+	if ip := net.ParseIP(conf.Ipaddr6); ip != nil {
+		if prefix, err := strconv.Atoi(conf.PrefixLen6); err == nil {
 			conf.Ipaddr6 = ip.String()
-		} else {
-			return fmt.Errorf("invalid ipv6 address: must use CIDR notation: %s", conf.Ipaddr6)
+			conf.PrefixLen6 = strconv.Itoa(prefix)
 		}
+	}
+
+	if ip, net, err := net.ParseCIDR(conf.Ipaddr6); err == nil {
+		conf.Ipaddr6 = ip.String()
+		prefixLen6, _ := net.Mask.Size()
+		conf.PrefixLen6 = strconv.Itoa(prefixLen6)
 	}
 
 	return nil
@@ -176,6 +177,28 @@ func (config *WarewulfYaml) IpCIDR() string {
 	cidr := net.IPNet{
 		IP:   net.ParseIP(config.Ipaddr),
 		Mask: net.IPMask(net.ParseIP(config.Netmask)),
+	}
+	if cidr.IP == nil || cidr.Mask == nil {
+		return ""
+	}
+	return cidr.String()
+}
+
+func (config *WarewulfYaml) IpCIDR6() string {
+	if config.Ipaddr6 == "" || config.PrefixLen6 == "" {
+		return ""
+	}
+	ip := net.ParseIP(config.Ipaddr6)
+	if ip == nil || ip.To4() != nil {
+		return ""
+	}
+	prefix, err := strconv.Atoi(config.PrefixLen6)
+	if err != nil {
+		return ""
+	}
+	cidr := net.IPNet{
+		IP:   ip,
+		Mask: net.CIDRMask(prefix, 128),
 	}
 	if cidr.IP == nil || cidr.Mask == nil {
 		return ""