+++ internal/pkg/config/root_test.go
@@ -153,6 +153,7 @@ ipaddr6: "2001:db8::1/64"
 `,
 			result: `
 ipaddr6: "2001:db8::1"
+prefixlen6: "64"
 warewulf:
   autobuild overlays: true
   grubboot: false
@@ -418,6 +419,59 @@ func TestNetworkCIDR(t *testing.T) {
 			conf.Network = tt.network
 			conf.Netmask = tt.netmask
 			assert.Equal(t, tt.cidr, conf.NetworkCIDR())
+		})
+	}
+}
+
+func TestIpCIDR6(t *testing.T) {
+	tests := map[string]struct {
+		ipaddr6    string
+		prefixlen6 string
+		cidr       string
+	}{
+		"blank": {
+			ipaddr6:    "",
+			prefixlen6: "",
+			cidr:       "",
+		},
+		"ip only": {
+			ipaddr6:    "2001:db8::1",
+			prefixlen6: "",
+			cidr:       "",
+		},
+		"prefix only": {
+			ipaddr6:    "",
+			prefixlen6: "64",
+			cidr:       "",
+		},
+		"full": {
+			ipaddr6:    "2001:db8::1",
+			prefixlen6: "64",
+			cidr:       "2001:db8::1/64",
+		},
+		"invalid ip": {
+			ipaddr6:    "asdf",
+			prefixlen6: "64",
+			cidr:       "",
+		},
+		"invalid prefix": {
+			ipaddr6:    "2001:db8::1",
+			prefixlen6: "asdf",
+			cidr:       "",
+		},
+		"ipv4 address": {
+			ipaddr6:    "192.168.0.1",
+			prefixlen6: "24",
+			cidr:       "",
+		},
+	}
+
+	for name, tt := range tests {
+		t.Run(name, func(t *testing.T) {
+			conf := New()
+			conf.Ipaddr6 = tt.ipaddr6
+			conf.PrefixLen6 = tt.prefixlen6
+			assert.Equal(t, tt.cidr, conf.IpCIDR6())
 		})
 	}
 }